Use lockable module to prevent race conditions in Batches.
Fixes #1362
Refs #1488

Ported from projecthydra-labs/curation_concerns#563 for Sufia 6.x implementations.
