+++ lib/sufia.rb
@@ -36,9 +36,17 @@ module Sufia
       #{Hydra::Engine.root}/app/models/concerns
     )
 
-    config.assets.paths << config.root.join('vendor', 'assets', 'fonts')
-    config.assets.precompile << /vjs\.(?:eot|ttf|woff)$/
-    config.assets.precompile << /fontawesome-webfont\.(?:svg|ttf|woff)$/
-    config.assets.precompile += %w( ZeroClipboard.swf )
+    initializer 'sufia.assets.precompile' do |app|
+      app.config.assets.paths << config.root.join('vendor', 'assets', 'fonts')
+      app.config.assets.paths << config.root.join('app', 'assets', 'images')
+      app.config.assets.paths << config.root.join('app', 'assets', 'images', 'blacklight')
+      app.config.assets.paths << config.root.join('app', 'assets', 'images', 'hydra')
+      app.config.assets.paths << config.root.join('app', 'assets', 'images', 'site_images')
+
+      app.config.assets.precompile << /vjs\.(?:eot|ttf|woff)$/
+      app.config.assets.precompile << /fontawesome-webfont\.(?:svg|ttf|woff)$/
+      app.config.assets.precompile += %w( ZeroClipboard.swf )
+      app.config.assets.precompile += %w(*.png *.jpg *.ico *.gif *.svg)
+    end
   end
 end
